+++ examples/magic-search/CHANGELOG.md
@@ -1,5 +1,12 @@
 # magic-search-demo
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @tollbit/magic-search@0.0.11
+
 ## 0.1.9
 
 ### Patch Changes